pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraudshield"
version = "0.1.0"
description = "Real-time fraud risk scoring with delayed labels and drift monitoring"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Abdul-Sobur Ayinde", email = "abdulsobur245@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core ML
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    
    # Data validation
    "pydantic>=2.0.0",
    "pandera>=0.17.0",
    
    # Monitoring & drift
    "scipy>=1.11.0",
    "evidently>=0.4.0",
    
    # Serving
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    
    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    
    # Metrics & reporting
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
fraudshield = "scripts.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "ml_regression: ML performance regression tests",
    "slow: Slow tests (skip with -m 'not slow')",
]
